+++ R/data2HTML.R
@@ -190,13 +190,6 @@ data2HTML=function(data,preprocessing="",path=".",filename="report.HTML",rawData
         mycat("\n\n")
 
     }
-    if(shiny::isRunning()){
-      progress <- shiny::Progress$new()
-      on.exit(progress$close())
-      progress$set(message = "Rendering File", value = 0)
-    } else{
-      cat("\nNow Rendering File:")
-    }
 
     if(type=="HTML"){
        out <- rmarkdown::render('report2.Rmd', rmarkdown::html_document())
@@ -207,14 +200,9 @@ data2HTML=function(data,preprocessing="",path=".",filename="report.HTML",rawData
     } else {
       out <- rmarkdown::render('report2.Rmd', rmarkdown::pdf_document())
     }
-    if(isRunning()){
-      progress$inc(1, detail = paste("Doing part", i))
-    } else(
-      cat("Done\n")
-    )
+
     result=file.rename(out, filename)
     #file.remove("report2.Rmd")
-
     setwd(owd)
     path=str_replace(path,"//","/")
     paste0(path,"/",filename)